requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
decouple==0.0.7
distro==1.9.0
Django==5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
multidict==6.0.5
openai==0.28.0
pydantic==2.6.4
pydantic_core==2.16.3
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
